refactor(singleton): 싱글톤 패키지 추가
