+++ src/main/java/lotto/Application.java
@@ -1,8 +1,6 @@
 package lotto;
 
-import lotto.model.lotto.Lottos;
-import lotto.view.InputView;
-import lotto.view.OutputView;
+import lotto.singleton.SingletonContainer;
 
 public class Application {
     public static void main(String[] args) {
