+++ src/main/java/lotto/singleton/Singleton.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.singleton;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
