+++ src/main/java/lotto/singleton/SingletonContainer.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.singleton;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
