+++ src/main/java/lotto/view/InputView.java
@@ -1,7 +1,7 @@
 package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
-import lotto.Singleton;
+import lotto.singleton.Singleton;
 
 @Singleton
 public class InputView {
